# backend/services/api-service/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Miniapp Platform API
  version: 1.0.0
  description: |
    API Gateway для платформы miniapps/webview.

    Основная логика:
    - Пользователь регистрируется только как user.
    - Admin может быть создан вручную в БД или назначен другим admin.
    - User видит опубликованные miniapps и может предлагать новые miniapps.
    - Admin управляет miniapps: publish, reject, edit, disable, enable, delete.
    - Запуск miniapp происходит через launch_token, который связывает user + miniapp.
    - Access JWT не передается внешнему miniapp.
servers:
  - url: http://localhost:8080
    description: Local api-service
...
security:
  - BearerAuth: []
paths:
...
  /api/ai/generate:
    post:
      tags:
        - AI
      summary: Generate miniapp from prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateMiniappRequest"
      responses:
        "201":
          description: Generated app and pending miniapp were created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedAppResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/ai/generated-apps/{generated_app_id}:
    get:
      tags:
        - AI
      summary: Get generated app metadata
      parameters:
        - $ref: "#/components/parameters/GeneratedAppIdParam"
      responses:
        "200":
          description: Generated app metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedAppResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/generated-apps/{generated_app_id}/config:
    get:
      tags:
        - AI
      summary: Get generated app config
      parameters:
        - $ref: "#/components/parameters/GeneratedAppIdParam"
      responses:
        "200":
          description: Generated app JSON config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedAppConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  parameters:
...
    GeneratedAppIdParam:
      name: generated_app_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    GenerateMiniappRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
          example: Сделай miniapp для учета расходов
    GeneratedAppResponse:
      type: object
      required:
        - generated_app_id
        - title
        - description
        - url
        - config
        - status
      properties:
        generated_app_id:
          type: string
          format: uuid
        miniapp_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          example: Expense Tracker
        description:
          type: string
          example: Miniapp for expenses
        url:
          type: string
          example: /generated-apps/1f5e3b2f-2f5f-4f8c-9d36-2e9fd1f8a2f4
        config:
          $ref: "#/components/schemas/GeneratedAppConfig"
        status:
          type: string
          example: active
    GeneratedAppConfig:
      type: object
      required:
        - title
        - description
        - pages
      properties:
        title:
          type: string
        description:
          type: string
        icon_prompt:
          type: string
        theme:
          type: string
        pages:
          type: array
          items:
            type: object
            required:
              - id
              - title
              - components
            properties:
              id:
                type: string
              title:
                type: string
              components:
                type: array
                items:
                  type: object
                  additionalProperties: true
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: bad_request
        message:
          type: string
          example: Invalid request body